新增IjkPlayer，支持IjkPlayer与MediaPlayer切换。

+++ nicevideoplayer/src/main/java/com/xiao/nicevideoplayer/NiceUtil.java
@@ -101,14 +101,14 @@ public class NiceUtil {
      * @param milliseconds 毫秒数
      * @return ##:##
      */
-    public static String formatTime(int milliseconds) {
+    public static String formatTime(long milliseconds) {
         if (milliseconds <= 0 || milliseconds >= 24 * 60 * 60 * 1000) {
             return "00:00";
         }
-        int totalSeconds = milliseconds / 1000;
-        int seconds = totalSeconds % 60;
-        int minutes = (totalSeconds / 60) % 60;
-        int hours = totalSeconds / 3600;
+        long totalSeconds = milliseconds / 1000;
+        long seconds = totalSeconds % 60;
+        long minutes = (totalSeconds / 60) % 60;
+        long hours = totalSeconds / 3600;
         StringBuilder stringBuilder = new StringBuilder();
         Formatter mFormatter = new Formatter(stringBuilder, Locale.getDefault());
         if (hours > 0) {
